added bin folder

+++ ArcaneEngine/src/ArcaneEngine/Core/Random.cpp
@@ -13,9 +13,9 @@ static uint32_t Lehmer32()
 	s_seed += 0xe120fc15;
 	uint64_t tmp;
 	tmp = (uint64_t)s_seed * 0x4a39b70d;
-	uint32_t m1 = (tmp >> 32) ^ tmp;
+	uint32_t m1 = static_cast<uint32_t>((tmp >> 32) ^ tmp);
 	tmp = (uint64_t)m1 * 0x12fad5c9;
-	uint32_t m2 = (tmp >> 32) ^ tmp;
+	uint32_t m2 = static_cast<uint32_t>((tmp >> 32) ^ tmp);
 	return m2;
 }
