+++ ArcaneEngine/src/ArcaneEngine/Graphics/ResourceCacheSource/ComputePipelineCreation.cpp
@@ -47,7 +47,7 @@ namespace Arc
 
         VkPipelineLayoutCreateInfo pipelineLayoutInfo = {};
         pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-        pipelineLayoutInfo.setLayoutCount = layouts.size();
+        pipelineLayoutInfo.setLayoutCount = static_cast<uint32_t>(layouts.size());
         pipelineLayoutInfo.pSetLayouts = layouts.data();
         if (desc.Shader->m_PushConstantSize > 0)
         {
